introduce ertm_nco_status
This (too) clever trick allows to alias the array of ertm_nco_reset
structures returned when getting status to a struct that accesses its
members as array or by lo/ref member name. There is no limit to the
amount of horrors that can happen during a morning coding bout.