pass c_char_p to ertm_init
The rationale for this is subtle: a char* in ctypes must be initialized
from a bytes, or None. So a bare None here would do. But, in the event
some poor soul changes None to a string, this will fail, because a str()
is a c_wchar_p, not a c_char_p. Hence, c_char_p('/dev/ttyUSB2') will,
at the very least, throw a run-time exception for the unwary.