WRS: make wrc compile for WRS
However, it does not fit the ram segment in memory.
In wrpc 4.2 the size was 23192 now is 29444 (probably can be reduced by 1KB
when different printf is used). The size of ram section in a switch was (based
on ld file) 0x6800 = 26624.
If ram segment is expanded then stats and mbox offets are different and
WRS software has to be adjusted accordingly.

Signed-off-by: Adam Wujek <adam.wujek@creotech.pl>