arch/lm32/ram-wrs.ld: expand ram segment
Since the code of wrpc used in WRS grew, the ram segment has to be expanded.
This required move of stats and mbox. Unfortunately, such move requires
a change of hardcoded offsets in WRS software. To avoid such needs in
the future stats and mbox were moved after stack. So in the future stack could
be reduced to make space for code.

This commit requires an adjustment in definitons of following in WRS software:
-- RTS_MAILBOX_ADDR in userspace/libwr/rt_client.c
-- FPGA_SPLL_STAT in softpll.h and in wrs_dump_shmem.c

Signed-off-by: Adam Wujek <adam.wujek@creotech.pl>