ertm14: enable CLK_EXT buffer on the ertm14 board
There's a bug on the PCB, the ext clock buffer enable is tied to EXT PLL (AD9516) reset, but the polarities differ. We don't need this PLL chip as the eRTM15 10M input offers superior timing performance. We just need the input buffer and the clk_10m_ext going to the FPGA for various calibration purposes.