ertm14: apply LPDC TX phase correction to PTP through SFP's deltaTX (WARNING, READ BELOW)
Note: I'm abusing the SFP deltaTX/deltaRX entries and just overriding them with bogus values adjusted only for the eRTM14's PHY TX asymmetry.

This is not the right way to handle the deltas. We need to separately provide the PHY delays and SFP delays to the PTP stack (and also probably drop the hardcoded bitslide parameter in ppsi). To be discussed.