Ensure functionality of WRPC testbench (testbench/wrc_core)
The following files are changed to ensure the functionality of the WRPC
testbench located in "testbench/wrc_core" folder of the wrpc-core repository:
- wrc_main_sim.c
- arch/risc-v/crt0.S
- arch/risc-v/crt0.h
- dev/endpoint.c
- configs/wrpc_sim_riscv_defconfig

Before compiling wrpc-sw for simulation, choose wrpc_sim_riscv_defconfig configuration.

In endpoint.c, delay function is commented out in the case of simulation.
In wrc_main_sim.c, passing information from the testbench to risc-v's software
is fixed and mini-NIC component is initialized correctly. Also, now the program
doesn't print the error message when the Rx FIFO of mini-NIC is empty, but
waits until there is something in the Rx FIFO, and then continues with checking
the correctness of the received frame.